Move the tests' tests into the same files and migrate UnitTests/variables.jl
This makes things much easier: you now write the test and a check with
MockOptimizer in a single place.

It's also a demonstration that this works for unit tests as well as the
larger integration tests.
