+++ src/Test/UnitTests/variables.jl
@@ -1,55 +1,37 @@
-#=
-    Functions in this file test functionality relating to variables in MOI.
-
-### Functionality currently tested
-    - add_variables
-    - add_variable
-    - deleting variables
-    - get/set VariableName
-    - is_valid for VariableIndex
-    - get VariableIndex by name
-    - NumberOfVariables
-
-### Functionality not yet tested
-    - VariablePrimalStart
-    - VariablePrimal
-    - ListOfVariableIndices
-=#
-
 """
-    add_variable(model::MOI.ModelLike, config::Config)
+    test_add_variable(model::MOI.ModelLike, config::Config)
 
 Test adding a single variable.
 """
-function add_variable(model::MOI.ModelLike, config::Config)
+function test_add_variable(model::MOI.ModelLike, ::Config)
     MOI.empty!(model)
     @test MOI.is_empty(model)
     @test MOI.get(model, MOI.NumberOfVariables()) == 0
     v = MOI.add_variable(model)
     @test MOI.get(model, MOI.NumberOfVariables()) == 1
+    return
 end
-unittests["add_variable"] = add_variable
 
 """
-    add_variables(model::MOI.ModelLike, config::Config)
+    test_add_variables(model::MOI.ModelLike, config::Config)
 
 Test adding multiple variables.
 """
-function add_variables(model::MOI.ModelLike, config::Config)
+function test_add_variables(model::MOI.ModelLike, ::Config)
     MOI.empty!(model)
     @test MOI.is_empty(model)
     @test MOI.get(model, MOI.NumberOfVariables()) == 0
     v = MOI.add_variables(model, 2)
     @test MOI.get(model, MOI.NumberOfVariables()) == 2
+    return
 end
-unittests["add_variables"] = add_variables
 
 """
-    delete_variable(model::MOI.ModelLike, config::Config)
+    test_delete_variable(model::MOI.ModelLike, config::Config)
 
 Tess adding, and then deleting, a single variable.
 """
-function delete_variable(model::MOI.ModelLike, config::Config)
+function test_delete_variable(model::MOI.ModelLike, ::Config)
     MOI.empty!(model)
     @test MOI.is_empty(model)
     @test MOI.get(model, MOI.NumberOfVariables()) == 0
@@ -57,15 +39,15 @@ function delete_variable(model::MOI.ModelLike, config::Config)
     @test MOI.get(model, MOI.NumberOfVariables()) == 1
     MOI.delete(model, v)
     @test MOI.get(model, MOI.NumberOfVariables()) == 0
+    return
 end
-unittests["delete_variable"] = delete_variable
 
 """
-    delete_variables(model::MOI.ModelLike, config::Config)
+    test_delete_variables(model::MOI.ModelLike, config::Config)
 
 Test adding, and then deleting, multiple variables.
 """
-function delete_variables(model::MOI.ModelLike, config::Config)
+function test_delete_variables(model::MOI.ModelLike, ::Config)
     MOI.empty!(model)
     @test MOI.is_empty(model)
     @test MOI.get(model, MOI.NumberOfVariables()) == 0
@@ -85,15 +67,15 @@ function delete_variables(model::MOI.ModelLike, config::Config)
     end
     @test !MOI.is_valid(model, v[1])
     @test MOI.is_valid(model, v[2])
+    return
 end
-unittests["delete_variables"] = delete_variable
 
 """
-    delete_nonnegative_variables(model::MOI.ModelLike, config::Config)
+    test_delete_nonnegative_variables(model::MOI.ModelLike, config::Config)
 
 Test adding, and then deleting, nonnegative variables.
 """
-function delete_nonnegative_variables(model::MOI.ModelLike, config::Config)
+function test_delete_nonnegative_variables(model::MOI.ModelLike, ::Config)
     MOI.empty!(model)
     @test MOI.is_empty(model)
     @test MOI.get(model, MOI.NumberOfVariables()) == 0
@@ -113,17 +95,17 @@ function delete_nonnegative_variables(model::MOI.ModelLike, config::Config)
     @test_throws MOI.InvalidIndex(v[1]) MOI.delete(model, v[1])
     @test !MOI.is_valid(model, cv)
     @test MOI.get(model, MOI.NumberOfVariables()) == 0
+    return
 end
-unittests["delete_nonnegative_variables"] = delete_nonnegative_variables
 
 """
-    update_dimension_nonnegative_variables(model::MOI.ModelLike, config::Config)
+    test_update_dimension_nonnegative_variables(model::MOI.ModelLike, ::Config)
 
 Test adding, and then deleting one by one, nonnegative variables.
 """
-function update_dimension_nonnegative_variables(
+function test_update_dimension_nonnegative_variables(
     model::MOI.ModelLike,
-    config::Config,
+    ::Config,
 )
     MOI.empty!(model)
     @test MOI.is_empty(model)
@@ -145,17 +127,18 @@ function update_dimension_nonnegative_variables(
     @test !MOI.is_valid(model, v[2])
     @test_throws MOI.InvalidIndex(v[2]) MOI.delete(model, v[2])
     @test !MOI.is_valid(model, cv)
+    return
 end
-unittests["update_dimension_nonnegative_variables"] =
-    update_dimension_nonnegative_variables
 
 """
-    delete_soc_variables(model::MOI.ModelLike, config::Config)
+    test_delete_soc_variables(model::MOI.ModelLike, config::Config)
 
 Test adding, and then deleting, second-order cone variables.
 """
-function delete_soc_variables(model::MOI.ModelLike, config::Config)
-    MOI.supports_add_constrained_variables(model, MOI.SecondOrderCone) || return
+function test_delete_soc_variables(model::MOI.ModelLike, config::Config)
+    if !MOI.supports_add_constrained_variables(model, MOI.SecondOrderCone)
+        return
+    end
     MOI.empty!(model)
     @test MOI.is_empty(model)
     @test MOI.get(model, MOI.NumberOfVariables()) == 0
@@ -171,15 +154,15 @@ function delete_soc_variables(model::MOI.ModelLike, config::Config)
     v, cv = MOI.add_constrained_variables(model, MOI.SecondOrderCone(3))
     @test MOI.get(model, MOI.NumberOfVariables()) == 3
     @test_throws MOI.DeleteNotAllowed MOI.delete(model, v[1])
+    return
 end
-unittests["delete_soc_variables"] = delete_soc_variables
 
 """
-    getvariable(model::MOI.ModelLike, config::Config)
+    test_get_variable(model::MOI.ModelLike, config::Config)
 
 Test getting variables by name.
 """
-function getvariable(model::MOI.ModelLike, config::Config)
+function test_get_variable_by_name(model::MOI.ModelLike, ::Config)
     MOI.empty!(model)
     MOIU.loadfromstring!(
         model,
@@ -193,15 +176,15 @@ function getvariable(model::MOI.ModelLike, config::Config)
     @test MOI.get(model, MOI.VariableIndex, "y") === nothing
     x = MOI.get(model, MOI.VariableIndex, "x")
     @test MOI.is_valid(model, x)
+    return
 end
-unittests["getvariable"] = getvariable
 
 """
-    variablenames(model::MOI.ModelLike, config::Config)
+    test_VariableName(model::MOI.ModelLike, config::Config)
 
 Test getting and setting variable names.
 """
-function variablenames(model::MOI.ModelLike, config::Config)
+function test_VariableName(model::MOI.ModelLike, ::Config)
     MOI.empty!(model)
     v = MOI.add_variable(model)
     @test MOI.get(model, MOI.VariableName(), v) == ""
@@ -213,17 +196,16 @@ function variablenames(model::MOI.ModelLike, config::Config)
     x = MOI.add_variable(model)
     MOI.set(model, MOI.VariableName(), x, "x")
     @test MOI.get(model, MOI.VariableName(), x) == "x"
+    return
 end
-unittests["variablenames"] = variablenames
 
 """
-    solve_with_upperbound(model::MOI.ModelLike, config::Config)
+    test_solve_with_upperbound(model::MOI.ModelLike, config::Config)
 
 Test setting the upper bound of a variable, confirm that it solves correctly,
 and if `config.duals=true`, check that the dual is computed correctly.
 """
-function solve_with_upperbound(model::MOI.ModelLike, config::Config)
-    atol, rtol = config.atol, config.rtol
+function test_solve_with_upperbound(model::MOI.ModelLike, config::Config)
     MOI.empty!(model)
     @test MOI.is_empty(model)
     MOIU.loadfromstring!(
@@ -244,7 +226,7 @@ function solve_with_upperbound(model::MOI.ModelLike, config::Config)
         x.value,
     )
     @test c2.value == x.value
-    return test_model_solution(
+    test_model_solution(
         model,
         config;
         objective_value = 2.0,
@@ -252,17 +234,36 @@ function solve_with_upperbound(model::MOI.ModelLike, config::Config)
         constraint_primal = [(c1, 1.0), (c2, 1.0)],
         constraint_dual = [(c1, -2.0), (c2, 0.0)],
     )
+    return
 end
-unittests["solve_with_upperbound"] = solve_with_upperbound
+
+function setup_test(
+    ::typeof(test_solve_with_upperbound),
+    model::MOIU.MockOptimizer,
+    ::Config,
+)
+    MOIU.set_mock_optimize!(
+        model,
+        (mock::MOIU.MockOptimizer) -> MOIU.mock_optimize!(
+            mock,
+            MOI.OPTIMAL,
+            (MOI.FEASIBLE_POINT, [1]),
+            MOI.FEASIBLE_POINT,
+            (MOI.SingleVariable, MOI.LessThan{Float64}) => [-2.0],
+            (MOI.SingleVariable, MOI.GreaterThan{Float64}) => [0.0],
+        ),
+    )
+    model.eval_variable_constraint_dual = false
+    return () -> model.eval_variable_constraint_dual = true
+end
 
 """
-    solve_with_lowerbound(model::MOI.ModelLike, config::Config)
+    test_solve_with_lowerbound(model::MOI.ModelLike, config::Config)
 
 Test setting the lower bound of a variable, confirm that it solves correctly,
 and if `config.duals=true`, check that the dual is computed correctly.
 """
-function solve_with_lowerbound(model::MOI.ModelLike, config::Config)
-    atol, rtol = config.atol, config.rtol
+function test_solve_with_lowerbound(model::MOI.ModelLike, config::Config)
     MOI.empty!(model)
     @test MOI.is_empty(model)
     MOIU.loadfromstring!(
@@ -281,7 +282,7 @@ function solve_with_lowerbound(model::MOI.ModelLike, config::Config)
     @test c1.value == x.value
     c2 = MOI.ConstraintIndex{MOI.SingleVariable,MOI.LessThan{Float64}}(x.value)
     @test c2.value == x.value
-    return test_model_solution(
+    test_model_solution(
         model,
         config;
         objective_value = 2.0,
@@ -289,94 +290,209 @@ function solve_with_lowerbound(model::MOI.ModelLike, config::Config)
         constraint_primal = [(c1, 1.0), (c2, 1.0)],
         constraint_dual = [(c1, 2.0), (c2, 0.0)],
     )
+    return
 end
-unittests["solve_with_lowerbound"] = solve_with_lowerbound
+
+function setup_test(
+    ::typeof(test_solve_with_lowerbound),
+    model::MOIU.MockOptimizer,
+    ::Config,
+)
+    MOIU.set_mock_optimize!(
+        model,
+        (mock::MOIU.MockOptimizer) -> MOIU.mock_optimize!(
+            mock,
+            MOI.OPTIMAL,
+            (MOI.FEASIBLE_POINT, [1]),
+            MOI.FEASIBLE_POINT,
+            (MOI.SingleVariable, MOI.GreaterThan{Float64}) => [2.0],
+            (MOI.SingleVariable, MOI.LessThan{Float64}) => [0.0],
+        ),
+    )
+    model.eval_variable_constraint_dual = false
+    return () -> model.eval_variable_constraint_dual = true
+end
 
 """
-    solve_integer_edge_cases(model::MOI.ModelLike, config::Config)
+    test_solve_Integer_with_lower_bound(
+        model::MOI.ModelLike,
+        config::Config,
+    )
 
-Test a variety of edge cases related to binary and integer variables.
+Test an integer variable with fractional lower bound.
 """
-function solve_integer_edge_cases(model::MOI.ModelLike, config::Config)
-    @testset "integer with lower bound" begin
-        MOI.empty!(model)
-        @test MOI.is_empty(model)
-        MOIU.loadfromstring!(
-            model,
-            """
-    variables: x
-    minobjective: 2.0x
-    x >= 1.5
-    x in Integer()
+function test_solve_Integer_with_lower_bound(
+    model::MOI.ModelLike,
+    config::Config,
+)
+    MOI.empty!(model)
+    @test MOI.is_empty(model)
+    MOIU.loadfromstring!(
+        model,
+        """
+variables: x
+minobjective: 2.0x
+x >= 1.5
+x in Integer()
 """,
-        )
-        x = MOI.get(model, MOI.VariableIndex, "x")
-        test_model_solution(
-            model,
-            config;
-            objective_value = 4.0,
-            variable_primal = [(x, 2.0)],
-        )
-    end
-    @testset "integer with upper bound" begin
-        MOI.empty!(model)
-        @test MOI.is_empty(model)
-        MOIU.loadfromstring!(
-            model,
-            """
-    variables: x
-    minobjective: -2.0x
-    x <= 1.5
-    x in Integer()
-""",
-        )
-        x = MOI.get(model, MOI.VariableIndex, "x")
-        test_model_solution(
-            model,
-            config;
-            objective_value = -2.0,
-            variable_primal = [(x, 1.0)],
-        )
-    end
-    @testset "binary with upper" begin
-        MOI.empty!(model)
-        @test MOI.is_empty(model)
-        MOIU.loadfromstring!(
-            model,
-            """
-    variables: x
-    minobjective: -2.0x
-    x <= 2.0
-    x in ZeroOne()
-""",
-        )
-        x = MOI.get(model, MOI.VariableIndex, "x")
-        test_model_solution(
-            model,
-            config;
-            objective_value = -2.0,
-            variable_primal = [(x, 1.0)],
-        )
-    end
-    @testset "binary with 0 upper" begin
-        MOI.empty!(model)
-        @test MOI.is_empty(model)
-        MOIU.loadfromstring!(
-            model,
-            """
-    variables: x
-    minobjective: 1.0x
-    x <= 0.0
-    x in ZeroOne()
-""",
-        )
-        x = MOI.get(model, MOI.VariableIndex, "x")
-        test_model_solution(
-            model,
-            config;
-            objective_value = 0.0,
-            variable_primal = [(x, 0.0)],
-        )
-    end
+    )
+    x = MOI.get(model, MOI.VariableIndex, "x")
+    test_model_solution(
+        model,
+        config;
+        objective_value = 4.0,
+        variable_primal = [(x, 2.0)],
+    )
+    return
 end
-unittests["solve_integer_edge_cases"] = solve_integer_edge_cases
+
+function setup_test(
+    ::typeof(test_solve_Integer_with_lower_bound),
+    model::MOIU.MockOptimizer,
+    ::Config,
+)
+    MOIU.set_mock_optimize!(
+        model,
+        (mock::MOIU.MockOptimizer) ->
+            MOIU.mock_optimize!(mock, MOI.OPTIMAL, (MOI.FEASIBLE_POINT, [2.0])),
+    )
+    return
+end
+
+"""
+    test_solve_Integer_with_upper_bound(
+        model::MOI.ModelLike,
+        config::Config,
+    )
+
+Test an integer variable with fractional upper bound.
+"""
+function test_solve_Integer_with_upper_bound(
+    model::MOI.ModelLike,
+    config::Config,
+)
+    MOI.empty!(model)
+    @test MOI.is_empty(model)
+    MOIU.loadfromstring!(
+        model,
+        """
+variables: x
+minobjective: -2.0x
+x <= 1.5
+x in Integer()
+""",
+    )
+    x = MOI.get(model, MOI.VariableIndex, "x")
+    test_model_solution(
+        model,
+        config;
+        objective_value = -2.0,
+        variable_primal = [(x, 1.0)],
+    )
+    return
+end
+
+function setup_test(
+    ::typeof(test_solve_Integer_with_upper_bound),
+    model::MOIU.MockOptimizer,
+    ::Config,
+)
+    MOIU.set_mock_optimize!(
+        model,
+        (mock::MOIU.MockOptimizer) ->
+            MOIU.mock_optimize!(mock, MOI.OPTIMAL, (MOI.FEASIBLE_POINT, [1.0])),
+    )
+    return
+end
+
+"""
+    test_solve_Integer_with_0_upper_bound(
+        model::MOI.ModelLike,
+        config::Config,
+    )
+
+Test a binary variable `<= 2`.
+"""
+function test_solve_ZeroOne_with_upper_bound(
+    model::MOI.ModelLike,
+    config::Config,
+)
+    MOI.empty!(model)
+    @test MOI.is_empty(model)
+    MOIU.loadfromstring!(
+        model,
+        """
+variables: x
+minobjective: -2.0x
+x <= 2.0
+x in ZeroOne()
+""",
+    )
+    x = MOI.get(model, MOI.VariableIndex, "x")
+    test_model_solution(
+        model,
+        config;
+        objective_value = -2.0,
+        variable_primal = [(x, 1.0)],
+    )
+    return
+end
+
+function setup_test(
+    ::typeof(test_solve_ZeroOne_with_upper_bound),
+    model::MOIU.MockOptimizer,
+    ::Config,
+)
+    MOIU.set_mock_optimize!(
+        model,
+        (mock::MOIU.MockOptimizer) ->
+            MOIU.mock_optimize!(mock, MOI.OPTIMAL, (MOI.FEASIBLE_POINT, [1.0])),
+    )
+    return
+end
+
+"""
+    test_solve_ZeroOne_with_0_upper_bound(
+        model::MOI.ModelLike,
+        config::Config,
+    )
+
+Test a binary variable `<= 0`.
+"""
+function test_solve_ZeroOne_with_0_upper_bound(
+    model::MOI.ModelLike,
+    config::Config,
+)
+    MOI.empty!(model)
+    @test MOI.is_empty(model)
+    MOIU.loadfromstring!(
+        model,
+        """
+variables: x
+minobjective: 1.0x
+x <= 0.0
+x in ZeroOne()
+""",
+    )
+    x = MOI.get(model, MOI.VariableIndex, "x")
+    test_model_solution(
+        model,
+        config;
+        objective_value = 0.0,
+        variable_primal = [(x, 0.0)],
+    )
+    return
+end
+
+function setup_test(
+    ::typeof(test_solve_ZeroOne_with_0_upper_bound),
+    model::MOIU.MockOptimizer,
+    ::Config,
+)
+    MOIU.set_mock_optimize!(
+        model,
+        (mock::MOIU.MockOptimizer) ->
+            MOIU.mock_optimize!(mock, MOI.OPTIMAL, (MOI.FEASIBLE_POINT, [0.0])),
+    )
+    return
+end
