+++ src/Test/contlinear.jl
@@ -495,6 +495,47 @@ function test_linear_integration(
     end
 end
 
+function setup_test(
+    ::typeof(test_linear_integration),
+    mock::MOIU.MockOptimizer,
+    ::Config,
+)
+    MOIU.set_mock_optimize!(
+        mock,
+        (mock::MOIU.MockOptimizer) -> MOIU.mock_optimize!(
+            mock,
+            [1, 0],
+            (MOI.ScalarAffineFunction{Float64}, MOI.LessThan{Float64}) =>
+                [-1],
+        ),
+        (mock::MOIU.MockOptimizer) -> MOIU.mock_optimize!(
+            mock,
+            [1, 0],
+            (MOI.ScalarAffineFunction{Float64}, MOI.LessThan{Float64}) =>
+                [-1],
+        ),
+        (mock::MOIU.MockOptimizer) -> MOIU.mock_optimize!(
+            mock,
+            [0, 0, 1],
+            (MOI.ScalarAffineFunction{Float64}, MOI.LessThan{Float64}) =>
+                [-2],
+        ),
+        (mock::MOIU.MockOptimizer) -> MOIU.mock_optimize!(mock, [-1, 0, 2]),
+        (mock::MOIU.MockOptimizer) -> MOIU.mock_optimize!(mock, [1, 0, 0]),
+        (mock::MOIU.MockOptimizer) -> MOIU.mock_optimize!(mock, [2, 0, 0]),
+        (mock::MOIU.MockOptimizer) -> MOIU.mock_optimize!(mock, [0, 2, 0]),
+        (mock::MOIU.MockOptimizer) -> MOIU.mock_optimize!(
+            mock,
+            [1, 1, 0],
+            (MOI.ScalarAffineFunction{Float64}, MOI.EqualTo{Float64}) =>
+                [-1.5],
+            (MOI.ScalarAffineFunction{Float64}, MOI.GreaterThan{Float64}) =>
+                [0.5],
+        ),
+    )
+    return
+end
+
 """
     test_linear_program(
         model::MOI.ModelLike,
@@ -600,6 +641,27 @@ function test_linear_integration_2(
             @test MOI.get(model, MOI.ConstraintBasisStatus(), c) == MOI.NONBASIC
         end
     end
+end
+
+function setup_test(
+    ::typeof(test_linear_integration_2),
+    mock::MOIU.MockOptimizer,
+    ::Config,
+)
+    MOIU.set_mock_optimize!(
+        mock,
+        (mock::MOIU.MockOptimizer) -> MOIU.mock_optimize!(
+            mock,
+            [1, 0],
+            (MOI.ScalarAffineFunction{Float64}, MOI.LessThan{Float64}) =>
+                [-1],
+            con_basis = [
+                (MOI.ScalarAffineFunction{Float64}, MOI.LessThan{Float64}) => [MOI.NONBASIC],
+            ],
+            var_basis = [MOI.BASIC, MOI.NONBASIC_AT_LOWER],
+        ),
+    )
+    return
 end
 
 """
@@ -734,6 +796,33 @@ function test_linear_inactive_bounds(
     end
 end
 
+function setup_test(
+    ::typeof(test_linear_inactive_bounds),
+    mock::MOIU.MockOptimizer,
+    ::Config,
+)
+    MOIU.set_mock_optimize!(
+        mock,
+        (mock::MOIU.MockOptimizer) -> MOIU.mock_optimize!(
+            mock,
+            [3],
+            con_basis = [
+                (MOI.ScalarAffineFunction{Float64}, MOI.GreaterThan{Float64}) => [MOI.NONBASIC],
+            ],
+            var_basis = [MOI.BASIC],
+        ),
+        (mock::MOIU.MockOptimizer) -> MOIU.mock_optimize!(
+            mock,
+            [0],
+            con_basis = [
+                (MOI.ScalarAffineFunction{Float64}, MOI.LessThan{Float64}) => [MOI.BASIC],
+            ],
+            var_basis = [MOI.NONBASIC_AT_UPPER],
+        ),
+    )
+    return
+end
+
 """
     test_linear_LessThan_and_GreaterThan(
         model::MOI.ModelLike,
@@ -822,6 +911,20 @@ function test_linear_LessThan_and_GreaterThan(
         @test MOI.get(model, MOI.VariablePrimal(), y) ≈ -T(100) atol = atol rtol =
             rtol
     end
+end
+
+function setup_test(
+    ::typeof(test_linear_LessThan_and_GreaterThan),
+    mock::MOIU.MockOptimizer,
+    ::Config,
+)
+    MOIU.set_mock_optimize!(
+        mock,
+        (mock::MOIU.MockOptimizer) -> MOIU.mock_optimize!(mock, [0, 0]),
+        (mock::MOIU.MockOptimizer) -> MOIU.mock_optimize!(mock, [100, 0]),
+        (mock::MOIU.MockOptimizer) -> MOIU.mock_optimize!(mock, [100, -100]),
+    )
+    return
 end
 
 """
@@ -987,6 +1090,21 @@ function test_linear_integration_modification(
     end
 end
 
+function setup_test(
+    ::typeof(test_linear_integration_modification),
+    mock::MOIU.MockOptimizer,
+    ::Config,
+)
+    MOIU.set_mock_optimize!(
+        mock,
+        (mock::MOIU.MockOptimizer) -> MOIU.mock_optimize!(mock, [4 / 3, 4 / 3]),
+        (mock::MOIU.MockOptimizer) -> MOIU.mock_optimize!(mock, [2, 0]),
+        (mock::MOIU.MockOptimizer) -> MOIU.mock_optimize!(mock, [4, 0]),
+        (mock::MOIU.MockOptimizer) -> MOIU.mock_optimize!(mock, [2]),
+    )
+    return
+end
+
 """
     test_linear_modify_GreaterThan_and_LessThan_constraints(
         model::MOI.ModelLike,
@@ -1098,6 +1216,20 @@ function test_linear_modify_GreaterThan_and_LessThan_constraints(
         @test MOI.get(model, MOI.VariablePrimal(), y) ≈ -T(100) atol = atol rtol =
             rtol
     end
+end
+
+function setup_test(
+    ::typeof(test_linear_modify_GreaterThan_and_LessThan_constraints),
+    mock::MOIU.MockOptimizer,
+    ::Config,
+)
+    MOIU.set_mock_optimize!(
+        mock,
+        (mock::MOIU.MockOptimizer) -> MOIU.mock_optimize!(mock, [0, 0]),
+        (mock::MOIU.MockOptimizer) -> MOIU.mock_optimize!(mock, [100, 0]),
+        (mock::MOIU.MockOptimizer) -> MOIU.mock_optimize!(mock, [100, -100]),
+    )
+    return
 end
 
 """
@@ -1256,6 +1388,20 @@ function test_linear_VectorAffineFunction(
     end
 end
 
+function setup_test(
+    ::typeof(test_linear_VectorAffineFunction),
+    mock::MOIU.MockOptimizer,
+    ::Config,
+)
+    MOIU.set_mock_optimize!(
+        mock,
+        (mock::MOIU.MockOptimizer) -> MOIU.mock_optimize!(mock, [0, 0]),
+        (mock::MOIU.MockOptimizer) -> MOIU.mock_optimize!(mock, [100, 0]),
+        (mock::MOIU.MockOptimizer) -> MOIU.mock_optimize!(mock, [100, -100]),
+    )
+    return
+end
+
 """
     test_linear_INFEASIBLE(
         model::MOI.ModelLike,
@@ -1345,6 +1491,31 @@ function test_linear_INFEASIBLE(
     end
 end
 
+function setup_test(
+    ::typeof(test_linear_INFEASIBLE),
+    mock::MOIU.MockOptimizer,
+    config::Config,
+)
+    if config.infeas_certificates
+        MOIU.set_mock_optimize!(
+            mock,
+            (mock::MOIU.MockOptimizer) -> MOIU.mock_optimize!(
+                mock,
+                MOI.INFEASIBLE,
+                tuple(),
+                (MOI.ScalarAffineFunction{Float64}, MOI.LessThan{Float64}) => [-1],
+            ),
+        )
+    else
+        MOIU.set_mock_optimize!(
+            mock,
+            (mock::MOIU.MockOptimizer) ->
+                MOIU.mock_optimize!(mock, MOI.INFEASIBLE),
+        )
+    end
+    return
+end
+
 """
     test_linear_DUAL_INFEASIBLE(
         model::MOI.ModelLike,
@@ -1423,6 +1594,24 @@ function test_linear_DUAL_INFEASIBLE(
             @test MOI.get(model, MOI.ResultCount()) == 0
         end
     end
+end
+
+function setup_test(
+    ::typeof(test_linear_DUAL_INFEASIBLE),
+    mock::MOIU.MockOptimizer,
+    config::Config,
+)
+    primal_status = if config.infeas_certificates
+        MOI.INFEASIBILITY_CERTIFICATE
+    else
+        MOI.NO_SOLUTION
+    end
+    MOIU.set_mock_optimize!(
+        mock,
+        (mock::MOIU.MockOptimizer) ->
+            MOIU.mock_optimize!(mock, MOI.DUAL_INFEASIBLE, primal_status),
+    )
+    return
 end
 
 """
@@ -1505,6 +1694,30 @@ function test_linear_DUAL_INFEASIBLE_2(
             @test MOI.get(model, MOI.ResultCount()) == 0
         end
     end
+end
+
+function setup_test(
+    ::typeof(test_linear_DUAL_INFEASIBLE_2),
+    mock::MOIU.MockOptimizer,
+    config::Config,
+)
+    if config.infeas_certificates
+        MOIU.set_mock_optimize!(
+            mock,
+            (mock::MOIU.MockOptimizer) -> MOIU.mock_optimize!(
+                mock,
+                MOI.DUAL_INFEASIBLE,
+                (MOI.INFEASIBILITY_CERTIFICATE, [1, 1]),
+            ),
+        )
+    else
+        MOIU.set_mock_optimize!(
+            mock,
+            (mock::MOIU.MockOptimizer) ->
+                MOIU.mock_optimize!(mock, MOI.DUAL_INFEASIBLE),
+        )
+    end
+    return
 end
 
 """
@@ -1615,6 +1828,26 @@ function test_linear_add_constraints(
                   MOI.NONBASIC
         end
     end
+end
+
+function setup_test(
+    ::typeof(test_linear_add_constraints),
+    mock::MOIU.MockOptimizer,
+    ::Config,
+)
+    MOIU.set_mock_optimize!(
+        mock,
+        (mock::MOIU.MockOptimizer) -> MOIU.mock_optimize!(
+            mock,
+            [650 / 11, 400 / 11],
+            con_basis = [
+                (MOI.ScalarAffineFunction{Float64}, MOI.LessThan{Float64}) => [MOI.NONBASIC, MOI.NONBASIC],
+                (MOI.ScalarAffineFunction{Float64}, MOI.GreaterThan{Float64}) => [MOI.BASIC],
+            ],
+            var_basis = [MOI.BASIC, MOI.BASIC],
+        ),
+    )
+    return
 end
 
 """
@@ -1808,6 +2041,57 @@ function test_linear_integration_Interval(
     end
 end
 
+function setup_test(
+    ::typeof(test_linear_integration_Interval),
+    mock::MOIU.MockOptimizer,
+    ::Config,
+)
+    MOIU.set_mock_optimize!(
+        mock,
+        (mock::MOIU.MockOptimizer) -> MOIU.mock_optimize!(
+            mock,
+            [5.0, 5.0],
+            con_basis = [
+                (MOI.ScalarAffineFunction{Float64}, MOI.Interval{Float64}) => [MOI.NONBASIC_AT_UPPER],
+            ],
+            var_basis = [MOI.BASIC, MOI.BASIC],
+            (MOI.ScalarAffineFunction{Float64}, MOI.Interval{Float64}) =>
+                [-1],
+        ),
+        (mock::MOIU.MockOptimizer) -> MOIU.mock_optimize!(
+            mock,
+            [2.5, 2.5],
+            con_basis = [
+                (MOI.ScalarAffineFunction{Float64}, MOI.Interval{Float64}) => [MOI.NONBASIC_AT_LOWER],
+            ],
+            var_basis = [MOI.BASIC, MOI.BASIC],
+            (MOI.ScalarAffineFunction{Float64}, MOI.Interval{Float64}) =>
+                [1],
+        ),
+        (mock::MOIU.MockOptimizer) -> MOIU.mock_optimize!(
+            mock,
+            [1.0, 1.0],
+            (MOI.ScalarAffineFunction{Float64}, MOI.Interval{Float64}) =>
+                [1],
+            con_basis = [
+                (MOI.ScalarAffineFunction{Float64}, MOI.Interval{Float64}) => [MOI.NONBASIC_AT_LOWER],
+            ],
+            var_basis = [MOI.BASIC, MOI.BASIC],
+        ),
+        (mock::MOIU.MockOptimizer) -> MOIU.mock_optimize!(
+            mock,
+            [6.0, 6.0],
+            (MOI.ScalarAffineFunction{Float64}, MOI.Interval{Float64}) =>
+                [-1],
+            con_basis = [
+                (MOI.ScalarAffineFunction{Float64}, MOI.Interval{Float64}) => [MOI.NONBASIC_AT_UPPER],
+            ],
+            var_basis = [MOI.BASIC, MOI.BASIC],
+        ),
+    )
+    return
+end
+
 """
     test_linear_Interval_inactive(
         model::MOI.ModelLike,
@@ -1898,6 +2182,27 @@ function test_linear_Interval_inactive(
             @test MOI.get(model, MOI.ConstraintBasisStatus(), c) == MOI.BASIC
         end
     end
+end
+
+function setup_test(
+    ::typeof(test_linear_Interval_inactive),
+    mock::MOIU.MockOptimizer,
+    ::Config,
+)
+    MOIU.set_mock_optimize!(
+        mock,
+        (mock::MOIU.MockOptimizer) -> MOIU.mock_optimize!(
+            mock,
+            [0.0, 0.0],
+            con_basis = [
+                (MOI.ScalarAffineFunction{Float64}, MOI.Interval{Float64}) => [MOI.BASIC],
+            ],
+            var_basis = [MOI.NONBASIC_AT_LOWER, MOI.NONBASIC_AT_LOWER],
+            (MOI.ScalarAffineFunction{Float64}, MOI.Interval{Float64}) =>
+                [0],
+        ),
+    )
+    return
 end
 
 """
@@ -2013,6 +2318,19 @@ function test_linear_transform(
     end
 end
 
+function setup_test(
+    ::typeof(test_linear_transform),
+    mock::MOIU.MockOptimizer,
+    ::Config,
+)
+    MOIU.set_mock_optimize!(
+        mock,
+        (mock::MOIU.MockOptimizer) -> MOIU.mock_optimize!(mock, [1.0, 1.0]),
+        (mock::MOIU.MockOptimizer) -> MOIU.mock_optimize!(mock, [0.5, 0.5]),
+    )
+    return
+end
+
 """
     test_linear_INFEASIBLE_2(
         model::MOI.ModelLike,
@@ -2111,6 +2429,31 @@ function test_linear_INFEASIBLE_2(
     end
 end
 
+function setup_test(
+    ::typeof(test_linear_INFEASIBLE_2),
+    mock::MOIU.MockOptimizer,
+    config::Config,
+)
+    if config.infeas_certificates
+        MOIU.set_mock_optimize!(
+            mock,
+            (mock::MOIU.MockOptimizer) -> MOIU.mock_optimize!(
+                mock,
+                MOI.INFEASIBLE,
+                tuple(),
+                (MOI.ScalarAffineFunction{Float64}, MOI.LessThan{Float64}) => [-1, -1],
+            ),
+        )
+    else
+        MOIU.set_mock_optimize!(
+            mock,
+            (mock::MOIU.MockOptimizer) ->
+                MOIU.mock_optimize!(mock, MOI.INFEASIBLE),
+        )
+    end
+    return
+end
+
 """
     test_linear_FEASIBILITY_SENSE(
         model::MOI.ModelLike,
@@ -2189,6 +2532,25 @@ function test_linear_FEASIBILITY_SENSE(
                 rtol
         end
     end
+end
+
+function setup_test(
+    ::typeof(test_linear_FEASIBILITY_SENSE),
+    mock::MOIU.MockOptimizer,
+    ::Config,
+)
+    MOIU.set_mock_optimize!(
+        mock,
+        (mock::MOIU.MockOptimizer) -> MOIU.mock_optimize!(
+            mock,
+            [1 / 5, 1 / 5],
+            (MOI.ScalarAffineFunction{Float64}, MOI.GreaterThan{Float64}) =>
+                [0],
+            (MOI.ScalarAffineFunction{Float64}, MOI.EqualTo{Float64}) =>
+                [0],
+        ),
+    )
+    return
 end
 
 """
@@ -2338,6 +2700,47 @@ function test_linear_integration_delete_variables(
     end
 end
 
+function setup_test(
+    ::typeof(test_linear_integration_delete_variables),
+    mock::MOIU.MockOptimizer,
+    ::Config,
+)
+    MOIU.set_mock_optimize!(
+        mock,
+        (mock::MOIU.MockOptimizer) -> MOIU.mock_optimize!(
+            mock,
+            [0, 1 / 2, 1],
+            con_basis = [
+                (MOI.ScalarAffineFunction{Float64}, MOI.LessThan{Float64}) => [MOI.NONBASIC],
+            ],
+            var_basis = [
+                MOI.NONBASIC_AT_LOWER,
+                MOI.BASIC,
+                MOI.NONBASIC_AT_UPPER,
+            ],
+            (MOI.ScalarAffineFunction{Float64}, MOI.LessThan{Float64}) =>
+                [-1],
+            (MOI.SingleVariable, MOI.GreaterThan{Float64}) => [2, 0, 0],
+            (MOI.SingleVariable, MOI.LessThan{Float64}) => [-2],
+        ),
+        (mock::MOIU.MockOptimizer) -> MOIU.mock_optimize!(
+            mock,
+            [1],
+            (MOI.ScalarAffineFunction{Float64}, MOI.LessThan{Float64}) =>
+                [-1],
+            (MOI.SingleVariable, MOI.GreaterThan{Float64}) => [0],
+        ),
+    )
+    # test_linear_integration_delete_variables has double variable bounds for
+    # the z variable
+    mock.eval_variable_constraint_dual = false
+    function reset_function()
+        mock.eval_variable_constraint_dual = true
+        return
+    end
+    return reset_function
+end
+
 """
     test_linear_VectorAffineFunction_empty_row(
         model::MOI.ModelLike,
@@ -2403,6 +2806,22 @@ function test_linear_VectorAffineFunction_empty_row(
     end
 end
 
+function setup_test(
+    ::typeof(test_linear_VectorAffineFunction_empty_row),
+    mock::MOIU.MockOptimizer,
+    ::Config,
+)
+    MOIU.set_mock_optimize!(
+        mock,
+        (mock::MOIU.MockOptimizer) -> MOIU.mock_optimize!(
+            mock,
+            [0.0],
+            (MOI.VectorAffineFunction{Float64}, MOI.Zeros) => [[0.0, 0.0]],
+        ),
+    )
+    return
+end
+
 """
     test_linear_VariablePrimalStart_partial(
         model::MOI.ModelLike,
@@ -2458,4 +2877,16 @@ function test_linear_VariablePrimalStart_partial(
         @test MOI.get(model, MOI.VariablePrimal(), y) ≈ zero(T) atol = atol rtol =
             rtol
     end
+end
+
+function setup_test(
+    ::typeof(test_linear_VariablePrimalStart_partial),
+    mock::MOIU.MockOptimizer,
+    ::Config,
+)
+    MOIU.set_mock_optimize!(
+        mock,
+        (mock::MOIU.MockOptimizer) -> MOIU.mock_optimize!(mock, [1.0, 0.0]),
+    )
+    return
 end
