+++ test/Bridges/lazy_bridge_optimizer.jl
@@ -1158,7 +1158,7 @@ end
         bridged,
         MOIT.Config{T}(solve = false),
         include = ["test_linear_"],
-        exclude = ["VariablePrimalStart"],
+        exclude = ["VariablePrimalStart", "linear_mixed_complementarity"],
     )
 end
